Override DEVELOPER_VERSION parameter so as to have unique ro.goo props
I cannot find where this is declared originally. It is set to false somewhere in my build process.
EDIT: I am keeping this as part f the process as reference, but setting this here does not affect the final build.prop, probably b/c of inheritance rules in makefiles. The next commit is what actually does the trick, in a hacky way.

Change-Id: If630d45c290ce404bf37b8ca238c198046bc4609